feat(progress): 进度条&环形进度条支持 changeData
- 默认开启 动画
